(test ver.) Fixed creation a lot of tiny DoubleArray
